I pretty much spent the entire today in thought.
I think the battle is coming down and the conclusion is that I need to redo the library from the ground up in order to get rid of generics. That means starting by making my own CudaDeviceVariable class that has no generic parameters.

The idea for this occurred to me first back in November I think, but I did not see the point of it then.

Back then I was writing about how the DU style might be ideal fit, but the reason why I am getting such poor composability and nongeneric code here is pretty much because F# is shit with generics and not even that great with tuples.

Considering my own Cuda backend now has the tuple map and F# does not, that is case in point.

One side benefit is that if I make my own CudaDeviceVariable class is that I will be able to support 16-bit floats as well.

Generics are just traps here to piss me off, they should just be left for containers and not touched for more serious things.

With that in mind, I'll forget about the map modules for a bit, while I deal with the very basics.

Once I am done with that, it will be a lot simpler to interface with backend and I won't have to deal with making a dozens of different function to interface with the backend like in my last two serious attempts.

It will not be efficient, but it will be one way of getting that well-typed API.