Before I can actually start on the library, I need to add two tedious features: Vg in the typechecker and SpiralCudaDeviceVar to replace CudaDeviceVar from ManagedCuda.
I wish it were otherwise, but as it is now, I will make the Spiral library as a dynamic interpreted language. And the big idea will be to augment it with dynamically scoped functions by passing them through the environment.

I'll use that in place of the template pattern.

Focus, focus. I need to put in just a little bit work today in order to get on the swing of things.

Let me do the two points above and after that, the rest won't be too hard. Only these two things will be tedious, the rest will be just fun.

Fun with finally finishing the library and going into training RL agents.