It does not feel like I did much today, but I did fill out like entire 12 pages in my journal, so I must have done something.
Today I've removed those process_tyvv and process_typedexpr functions and moved all the destructuring functionality to the destructure function.

As far as I am concerned, this is the way to pattern matching tuples. It used to be a large issue for me, but now is completely trivial.

The important thing is that I have a firm grasp of what pattern matching does so the firm separation of functionality that exist in the code is the same as it exists in my head.

...There are a raft of other improvement that I do not feel like getting into here as I am tired.

Suffice to say, I am nearly done with the codegen and I will get these kernels to compile tomorrow. After that I will make the interfaces for the modules.