I fixed the code generator and figuring out that the reality algorithm was not correct, changed it significantly simplyfing it in the process.
Not only is it more correct now, but it is significantly more roboust as there is only one global stack whereas before each method had its own.

I am not sure if it completely correct, but it might be and this is a step in the right direction.

Last night I also figured a better way of doing memoization after which I implement it, the typechecker will trully be perfected.

I did not expect there to be a better way of doing recursive inference and neither did I expect this better way of naming things, but I did.

For now let me busy myself with putting this in. After that I will figure out how to deal with kernel naming and hashing.

Right now it compiles, but since the leader requires a specific name for the main kernel, I need to figure out how to deal with this.