I managed to extend the typechecker with custom memoization. I definitely needed this, but I have doubts with the way code is organized now.
I feel that method memoization and destructuring should not be separate and neither should the closure conversion.

I am thinking that I compiled with CPS, I could get some radical simplification here, but I am not sure whether that would be doable while compiling to C++.

It is just a thought and I do not see how one could merge memoization and destructuring.

Also currently there is a bug in that duplicate arguments will give an error in function calls, that I do not feel like fixing.

...Oh, well, it can't be helped that first attempts at doing a thing turn out to be overly convoluted.

Eventually, I will have to tear what I have built here down and redo it from scratch, but now I am actually proud of what I have now.

For making a bunch of generic map modules, I really went beyond the call of duty here. Before I call it done, maybe I will improve tuple destructuring so indexing is not rebound...

But that would create problems with printing of variables which is linked to the aforementioned bugs.

Agh, seriously, there is so much trouble around function boundaries, it is ridiculous. This can't be really it, but it is good enough for now, until I come up with a new way of thinking about compilation.

The next thing on the agenda should be to finish the loader for the modules and then to get to work on the library.

I'll take it a bit easier from here; yesterday I was fully in the zone the whole day and I need a bit more leisurely pace.