No, it is no good. I am going to backtrack. For reasons why check out the comment at the top of `Typechecker_v7a.fsx`.
It is a good thing I had the foresight to anticipate that implementing this might not work out and opened up new files.

With that, I'll refactor the 6e a bit and move forward with that then.

First let me rest a bit.

I put in a lot of effort into making those things yesterday, but I could do that because I spent so much time thinking about it first.

In either case, for a certain definition of done - I am in fact done with the backend and by extension with the rest of the library as well.

Making the typechecker was very much a new thing for me and required a lot of creativity on my end, but apart from name hashing which won't be hard, I do not think the rest of the library will be particularly challenging.

At least not compared to this.

I just need to roll up my sleeves and work it out.

Yeah, I am done, I am very nearly done.

I think I will take it easy until the time comes to venture into RL once more.

No doubt the skills I've acquired over the past two years will come in handy.