Done tidying things up.
Actually, forgot that before I can do anything else, I need to make the dynamic CudaDeviceVariable. This will be some work. I'll have to copy the code in ManagedCuda's source. Last time I looked it was over 1k lines long so I do not feel like starting this right now as it is late.

I'll call it a day with this.

I know that I did the reverse of work today. Forget about resting. Trying to rest just make me even more anxious. I will have plenty of time to rest while my agents are training.

Despite everything it won't be hard to finish the library. The typechecker itself was harder than it.

It will go by quickly, I am sure of it.