Ok, enough. I am definitely not starting a new round now.
It might seem straightforward to just start putting in things now that I have the new device variables, but I can't really tell what was important in the DM file and what is not. I am also thinking whether I should change the allocation scheme, but forget that for now.

Tomorrow on the agenda it will be to bring back the ObjectPool and after that it will be time to start work on the primitives.