I had a great leg up today having woken up early and slept well, but rather than work on the DM class, I spent most of this time thinking about CPS and closure conversion.
Right now the function for closure conversion (it would be better called implicit argument propagation) is only half complete. I know how to do the other half, but...

I think I did a decent job with the typechecker, but it is not enough.

The type inference aspect of it is done with and I am not interested anymore in that aspect of it, but I am unsure whether the representation I've chosen as the base is optimal.

I still do not really understand CPS compilation.

Like ok, you need to call the closure at the end and the methods all have void return types, but what do you do for example with tuple expressions, that is if every expression needs a continuation? The stuff in the tuple obviously does not continue anywhere.

In theory if I could turn everything into a function CPS'd call, not only would I not need separate cases in destructuring for tuples, but in addition to that, maybe the bound previous, used up half in the cc function would be enough for everything.

And it would open the door to fun stuff such as effect handlers and maybe more.

Right now instead of passing higher order function around, it would be more precise to say that I am passing them around at the type level. I think the same could be done with effect handlers.

...No, the point is not about effect handlers.

To finish the library, all I would need to do is to finish the DM class and the rest, but…

Since the Spiral library is its own language, not internalizing the correct design principles will have a negative effect through each and every line I write from here.

I can't cheat this. It simply does not work like that. Programming is not like that. You do not do things incorrectly and then come back to do it right. If I did that, it would be essentially the same as saying that it is okay that I program badly. It would be saying that I am okay with being a bad programmer for some kind of near term reward.

Writing bad code when you don't know what good is cannot be helped, but the programmer's intuition inside me is clear.

In terms of writing languages there is one level yet above me. If I can reach it, then I will be able to make the Spiral library truly spectacular.

Just like I did in February, I think I will take a few days off to think and after than maybe a few weeks to see if I can find a new path for this.

...Maybe. As a preliminary, let me try Appel's Compiling With Continuations again and then I will decide what to do after that.