Actually, let me do it now.
Today I manged to finish the Compilation with CPS book by Appel and then did some refactoring and bug fixing. I also implemented the free variable algorithm correctly based on what I saw in the book.

I did not get much out of what I wanted, but unlike in February when I first got the book and just skimmed it, this time I feel like I learned quite a lot.

I am not sure what I expected, but I did not expect CPS compilation to be the lambda calculus version of assembly.

What I really wanted to find in fhe book was resolve some of the issues with closure passing that have been plaguing me, and with the book as a catalyst, I managed to think them through and realized that there is nothing wrong with the stack-allocated kind of closures that I am using now.

In fact the book talked about downward and upwards passing closures and it made me decide to not support the upwards passing (or function returning) of closures. Implicit argument propagation of the kind I am doing now is enough.

And today I baked in a few optimization into destructuring so no unnecessary intermediate variable creation is made during destructuring of tuple indexing expressions.

The bug mentioned that I fixed is during passing of literals. I did not do it right and have even considering passing them via the type system as dependent types, but I decided against it.

Unless I have some pressing need for metaprogramming, there is no need to bake in more type system features.

Now that everything is in place, I am actually very, very pleased with how the DSL turned out.

It is pretty much perfect for what I set out to make out of it.

Reading the CPS book by Appel gave me an alternative view in what a real compiler does and with that, a second hand perspective on what I should expect from a DSL.

Two days ago, I did not really understand what CPS compilation was about and had various misconceptions - it is definitely a different thing that programming in the CPS style itself.

Yesterday it seems to me that CPS is half of the key to the puzzle of programming; I am amazed that those particular set of restrictions allow lambda calculus to essentially act as assembly.

The way algebraic datatypes were structured was quite instructive and there is no way I could have figured out what the constraints on the language were from those Lisp tutorials on CPS.

With this I am ready to start work on the Spiral library proper.

I've considered reading Harper's book and that dependently typed programming course in Scala that got posted on the /r/scala sub this morning, but I had enough of math.

Based on what I learned on putting in dependent types by putting them into the Cuda DSL, I feel like I know all I need for now. I am satisfied with my current level of knowledge.

I guess it is time for more programming after all.

This time I really will finish the library and go into reinforcement learning.