Working on APIs is surpringly tedious.
At any rate, I am done with the big 3: geam, gemm and saxpy.

To be honest, it is kind of hard going.

Six months ago, I thought F#'s type inference is great, but now I think it is actually pretty shitty compared to what it could be and its polymorphism support is very minimal.

Well, I'll see how it goes as I move through this. In the worst case, I'll write a DSL for just the API if things starts getting annoying.