I went to remove duplicate filtering because It would interfere with outside methods and I discovered a bug in the printing of implicit arguments.
I did quite a good job finding it since the changes I introduced into the code were not related to it at all. I really was a ninja this time; if it was the old me, it would have taken me a lot longer to find this.

Apart from that, I've decided what I am going to do with the library.

I am going to make it a DSL after all, and make a specialized typechecking pass for it. The reason for that is so I can memoize kernel creation and not have to deal with hoisting them like I did in the last iteration of the library.

I've been thinking that it might be possible to do without a typechecking pass, but I've been making that same mistake for the past six month and I won't do it again.

If I want generic code with any kind of performance, typechecking is necessary.

I'll get right onto it.