I am just starting now.
Before I get cracking let me just state that I am a bit taken aback at this. Yeah, I decided to do typechecking, but I had not decided yet whether to put in variables, functions and everything else.

And once I have variables and functions, I essentially have a full language right there.

I've decided I need it, so I will do it.

That means that I will essentially be reimplementing the system for the backend here...

Not what I had in mind.

As an aside, the Idris book was released 10 days ago and I am currently waiting for it to become available on library Genesis. If I had a language with more advanced types, maybe I could have avoided all this work, but there is no helping it.

If the AD language ends up being a copy of the backend maybe I will have to take a look into unifying them.

As it is, I am going to have to endure some code duplication here.

Making that backend was hardly a piece of cake - if I had to do both the main language and it under one umbrella I'd never find the focus to get it off the ground.

Let me do it bit by bit with the pieces that I have now.