I am pretty much dry today. I put in zero forethought in how I am going to do the typechecking for the main library and only decided that I am going to do typechecking at all a few hours ago.
I intended to do it separately, but as soon as I started to write it occurred to me just how much in common the two typecheckers would have.

There is no way I am reimplementing all this in different place, nor am I copy pasting this. I will have to merge the backend Cuda language with the main one.

But most of my programming day is already over and I do not feel like starting now, so I guess I will take off an hour and a half earlier than usual.

Lying to myself that it will be easy hardly ever works out, but it gets me to make the first step.

Making the library is difficult for me because I am still tied to the old and busted patterns from the previous iteration.

But unifying the two languages into one might be ideal - I will finally abstract that shitty GPU out way beyond the border of my thoughts.

This is the way it should be.

Instead of two separate pieces, what I should have imagined is a hierarchy.