"Yesterday when I got up I was pretty much dead tired already, and writing that grant thing did not help. Writing the About Me just made me realize how fucked I am unless I can get that one win. There is no way I'd get the grant...unless I get that one win first. And if I get that win, the grant will not matter even in the slightness. How shameful weakness is.
Well, forget about the grant thing, I have a lot bigger problems than it.

Here is one of the plans I came up with:

- First get all the code out of the type checker. Pull it into functions and replace the typechecker cases with function calls. That will allow me to partition the language.

- Go with the segregation idea. Reuse the code from the typechecker. It will be easier now that it is partitioned.

- Implement the second sublanguage…

This third point is the sticker. The above plan sounds great on the surface and it did for me until I thought a bit further ahead.

Just how the hell am I going to implement the layers? Obviously, the same way I do it in standard F#. For that I need the option type.

For that I need algebraic datatypes with existential types and pattern matching on them.

So then I would need to implement that.

The bad news just keep on rolling. Imagine just how it would be to program in such a language. Take the syntax from the Cuda modules code and extrapolate it for thousands of lines…

Ugh. There is no way around it; in addition to algebraic datatypes and complete pattern matching, I'd have to invent a new syntax and implement a parser for that.

At that point, the Spiral library is not a library, or even a DSL – it would be a fully fledged language!

This can't be happening, just what did I do to deserve this? I can't take responsibility for this, all I really wanted was to memoize those map modules and make calling them fast. Just why the hell would I have to implement a full language for that?

---

Alternatively – just use a dictionary to memoize them and write the library in plain F# as intended.

This won't be as fast I want it to be, but it certainly faster than doing all of the above.

Maybe the speed of calling the kernels will not be that horrible?"

The above is from my journal.

Here is what I wrote on the F# sub.

https://www.reddit.com/r/fsharp/comments/652sin/what_are_the_worst_parts_of_f/dg7awcc/?st=j1g4wnyk&sh=7b19d4d4

So for the past two hours since I got up I've been basically doing this. I will go forward with the memoization approach. I do not feel like doing a fully fledged language right now.

I'd at least want to go through the Idris book and get some dependently typed programming experience first.

To be honest, I am nearly ready to be converted.

You know, when I saw the Grenade library in Haskell, I thought it was pretty ridiculous and wasteful to pass layer information through the type system, but, BUT...if I had that type driven metaprogramming capability I could memoize the kernels without having to implement an entirely new language!

I do not think Haskell would be actually able to do this, not without breaking out the Template Haskell aspect of it, but it is just that passing stuff like functions and data through the type system makes a lot more sense in this interoprability scenario.

...

Well, there is no going around this. I am simply going to have to let luck bail me out here and go with memoization using global dictionaries approach. The GPU launches being asynchronous is one silver lining in all of this.

Yesterday was one of those days, I must have gone to bed at 6:30pm and had not even bothered to brush my teeth.

At least I am rested today, yesterday I had a nightmare and woke up exhausted.

From the looks of things, doing every correctly will be impossible, I'll push of the issues of integrating the backend into the library proper a few years into the future when I will have the tools to deal with this.

For now, I will take the win that I have and just be happy with what I did since the start of the year.

I went way beyond the call of duty with this Cuda backend.