I did not do any coding today because I was busy trying to figure out what is Justice.
I know now.

I think that my resolution for 2018 should be to ditch F# at that point and pick up a dependently typed language like Idris or do my own language. As it is now, F# is simply too inexpressive.

Right now I feel the same as in January when I was depressed that I cannot generate Cuda code directly.

I mean, of course I do not want DM and D to be dynamic! It is an embarrassment to me to have the library be like this.

Despite all the past effort I put into the library when V5 is finished, I'd barely give it a passing grade.

The syntax to access the map modules will be horid, I will still have to take care to hoist them out of functions manually and the calls to the backend generated function will be slow. Despite being an advanced functional language, the library will feel as if it was written in a dynamic one with all the type unsafety that implies.

It is no wonder why my body felt like lead yesterday and that I can barely write a line of code.

Credit where it is due, unlike before when the Cuda segment of the library was by far the weakest part, right now the backend is perfect. It is by far the best part of Spiral. It literally changed my mind on the value of dependent types. I've been way too focused on the needless theorem proving aspect of that.

It is the rest of the language that is the problem now.

With this new backend, I will be able to write safe generic code, so with that I won't have to worry anymore about passing wrong arguments and missing an edit because I copy pasted.

I thought ensuring that safety and expressiveness was absolutely necessary so I put in two months to work on the backend and now it is done.

With that the library meets the minimum criteria for usability in my eyes.

But it could be so much more if it was written in a better language and I did say that I was going to make my own language eventually.

And I am going to have to learn dependently typed programming before that.

When I wrote that depressing grant two days ago, I was actually trying to get money for this library, the one that will be finished in two weeks or so.

Regardless how it went, I doubt that this thing deserves it. It is barely passing and I can use it finally go into RL as I intended, but if somebody gave me 5k to rewrite the library in Idris that would be quite an interesting offer indeed.

So that is what the grant should be about.

Right now there is not a single library with a well integrated GPU backend and there won't be simply because none of the languages save for Idris and other dependently typed ones have the tools necessary for proper interoperability.

I certainly won't be able to do it in F#.

If I resolve myself to do this, I could probably get that grant, but I would have to dedicate a good chunk of my time and delay my entry into RL once again.

But it would definitely be worth it. If I rewrote Spiral in Idris, I have what it takes to make it the best DL library in the world. I wouldn't have been able to do it six months ago, but I think a challenge of such caliber is right up my alley now.

And by the end of I would have that great library I've been dreaming of for the past 1.5 years and I would learn the very frontier of programming. At that point I would become a very scary opponent to my future enemies.

And if by some change I fail to get the grant, then nothing to it. I'll have finished V5 by the end of the month and just proceed with the plan I have already. I'll pick up dependently typed programming eventually after I've made some money raiding the online gambling sites.

I'll sit on this idea for a week to see if I change my mind because spending like six month rewriting the library in Idris is a potentially long time and I intend to stick to the rules of the grant.

As I am dead tired, mostly from stress, I'll call it a day here and start work on the library proper tomorrow. Hopefully I'll have finally gathered my will by then.

This code is incredibly difficult for me to write because I know it is suboptimal and yet there is absolutely nothing I can do about it. There is no way I would ever tolerate such shoddy quality in my own language.