I did some major refactoring in order to make the DM types mesh better with the typechecker.
I am a bit exhausted from this coding spurt. I need to rest a bit before continuing.

I need to check whether my rewriting broke the typechecker or the codegen and whether this thing works. Then I will time the complete kernels and compare it to this.

Right now I feel like I am getting up from a long slumber. At the very least I am finally drumming out code.

That map_op bit is pretty much the reason I put in MapVV. Thanks to that wrapping, the function passed in will expect normal inputs and be expected to return a tuple which will then get fed to the outputs.

It is truly a beautiful thing.