I did various things including fixing the type inference algorithm for recursive cases. I forgot how it worked and in the process of relearning it, I realized there were two things wrong with it.
Last time I was not sure, but now I am decently sure that it is correct. The fix I did was actually pretty inspired, there is no chance of infinite loops anymore.

Also I put duplicate name filtering back in and modified CubBlockReduce so it works in all cases.

There is no doubt about it, right now it is still a hack and when the time comes to add complete pattern matching to the language I am going to have to rewrite this. But it is good enough for now.

I also fixed a bug in destructuring. The way it was before, it would not do deep destructuring on tuples.

It seems that now that I am in my normal frame of mind, the improvements just keep coming.

I am going to have to cut my coding day short as I nee to help my father deal with SSL certificates in Linux. I do not know anything about that and if experience serves, it is probably going to be a pain in the ass and take a while.