It seems I forgot to push this an hour ago because the Internet was wonky.
At any rate, I am done with complete pattern matching on tuples. Now with this, I will be able to treat them as heterogenous lists.

I am also a small step away from adding pattern matching on open types.

Damn, I am tired. In terms of absolute lines of code written, the scope of the changes was small, but damn did it take it out of me.

This feeling of giving it all is like I am literally channeling my will into the code.

Looking back at it, perhaps the most personal reason of why disbelieved all those retarded paperclipper AI scenarios is because of this single thing.

Programming is a release of will.