Today I had some back pain so I had to exercise to get it out of my system. Probably, I had a really bad posture while programming yesterday and I had trouble sleeping because of it.
I am treading warrily because of that today.

I definitely did not take it easy yesterday and am paying the price now.

At any rate, I really had a lot of trouble trying to wrap my head around on how to do pattern matching on tuples in its full generality.

I said I was done yesterday, but it was not even close. A simple thing like this turns out to be qutie difficult.

And speaking of that, a simple task like zipping a tuple is something I have absolutely no idea how to do.

I actually can't think of a method to do it.

It will take me a while to figure out. But first, let me rest a bit.