I knew that zipping arbitrary tuples would be a hard task and I see that my expectations were not wrong. I am still not done with it yet, but I had enough for today.
Since fully generalized zipping is a multidimensional transpose, it is no wonder it is taking me so long to do this.

The fact that I am not transposing pure lists just makes things harder.

Ugh, this is such a pain in the ass. Hopefully I will be done with this tomorrow.