I finally managed to do the generalized tuple zip in CPS form. This trivial thing of tuple zipping is already 130 LOC long and took me most of the day.
I am not sure I have the energy for any other big projects today.

Now that I got some experience writing programs like this, I think I can rewrite the pattern matcher in this form and make it a bit more efficient.

Two days ago, I was playing with the idea of writting the zip in the language directly, but this is quite difficult. I'll buld it in on the F# side.

It would be a nightmare to do it with current syntax and no editor support.

Edit (4h later): It seems I forgot to push this after writing it.