This third version is perfect. It has to be optimal to do zipping. there is no doubt in my mind about it.
To be honest, it is earlier than usual, but I had enough for the day. It is just late enough that I do not feel like starting a new project since I only have an hour left until I am out of steam. I'll rewrite the pattern matcher in the CPS form tomorrow and put in the VVZip case based on what I have written today.

I pretty much spent the entirety of today just optimizing the zip function. When I first started it yesterday, I actually did not have any idea it was actually a dimensional shuffle operation or how to do it in a generic manner.

I am proud of this code, I definitely would not be ashamed to show v3 to anybody else. This is functional programming in its essence.