Last night I did a lot of thinking on how to typecheck the CPS scenario and figured out a few flaws in my current scheme. Putting the raw Env directly into the type was horrible since every variable would have an unique tag. Any kind of recursive scheme would cause the typechecker to diverge.
I fixed this so the types now take enviroments to types and that should be it. This thing with types is really making me understand how shitty dependent typing is for typechecking especially when it comes to tracking variables at the term level. If it was not need for that, it would be way easier to do it.

But assuming there aren't any dependent arrays, now this scheme should suffice for any kind of recursive typechecking scenario. I can write my code in CPS without worries knowing it will complete.

With this I feel the typechecker's general scheme for type inference for function is now complete. I doubt there will be any more changes on this front.

It is amazing how much effort this one thing took. It was long in coming.