I had enough for the day.
I've done a significant amount of work on mapping over tuples today and made quite a few syntactic sugar functions.

Programming in the backend should be quite a bit more pleasant going forward.

I haven't tested this yet. I am saving that tomorrow.

Tomorrow, I will tests all of this and make that generic AtomicAdd in the vein of Mset.

Today, I was not particularly inspired, but instead just made slow and steady progress.

I just need to keep going. Eventually I am going to finish this thing and be in the correct position when the situation starts to heat up.