I admit I slacked a bit in the last hour since I downloaded Bayonetta. It got released a week ago on the PC. Actually, I really with Denuvo would get cracked so I can play Nier Automata...
Right now I have a bit of a problem in the typechecker. There is a bit of boilerplate I want to abstract over, but to do it, I would either need Lisp styled unhigienic macros, or to make the typechecker more like a parser combinator with the ability to backtrack pattern matching cases.

To do that I'd either need to make everything immutable and use the state monad pattern or I will just have to make deep copies at certain points.

Deep copies would probably be easier, not to mention more performance efficient.

Agh...let me just do it.