I did my duty for day.
Just an hour ago, doing this kind of extensions to the type checker seemed like it would be huge effort and require a lot of boilerplate, but what happened is that I wrote just a few lines and with that pretty much changed the nature of the typechecker.

This is functional programming for me, it exceeds my expectations time and time again.

I was unwilling at first, and yet once I got the desired insight I wrote some of my best work to date.

CPS is wonderful as well.

I am still not done yet though. I need to write the case_f function and I am not sure where to do it. Probably at around apply_template. Maybe I will make a recursive block there.