This is a failure. I am just wasting my time.
Basically, rewriting the pattern matcher to support evaluation is simply too difficult. I can't do it.

The reason for that is because it is in CPS form.

It is fine for trivial problems like the zip function, but ultimately CPS is just a convention and I am not getting any help from the type system to ensure CPSed ness here.

Trying to rewrite the typechecker body to support it is simply hell.

I am getting those ridiculous type errors with deeply interwoved lambdas like I did on the failed third attempt of the compiler in January.

I am unrolling them manually and simply bashing my head against the wall here.

I am an idiot. Passing along failure cases as higher order function is a great idea, but going a step further and doing away with the main stack is a step too far.

I'd be better off using the Result type instead.

Even parser combinators, whose pattern I am trying to emulate here do the same thing - FParsec uses the Reply type for the return from its parsers.

Let me clear my head for a while and then I will try to do this again. I will remove CPS from the typechecker, including the zip functions and do it from the start using the Result type for parts of the code that would ordinarily require CPS.

That would be for the best. I honestly doubt that allocating a closure every time I have to make a function call will be that much faster than allocating a DU anyway and it certainly would not be if the DUs were value types.

And after I do that, I won't have to deal with these convoluted type errors or twist my mind because I suddenly do not have a stack anymore.