Let me push this here.
That it took me so long to find these type errors is a bit demoralizing. Once I converted the thing to use the Result type, it was an easy fix.

It is rare that I spend an entire day debugging, and usually when that happens to me it is due to a type error.

So what do I do here?

As I said, I was really close to solving this, I just lost track of things in my head, or rather lost confidence in myself. But as I compare the two versions, I can see that they are very close to each other.

Yeah, I've all but decided that algebraic datatypes are the way to go, but there is just one snag that I've realized only now.

It does not matter since that is the way type inference is done for ML languages, but algebraic datatypes would require unification to infer the types of their generic parameters.

This is not something I will be able to do in Spiral without a significant bit of complexity.

I am going to have to use higher order functions where here I would use DUs.

Well, let me just continue working on this and I will decide at the end, which version do I like better.

It might be the case that I simply need more experience with CPS.