I converted the typechecker so it uses the new system. Good.
Before I do anything else, if I am going to use this style I need to linger on my mistake and analyze it in order to make sure it does not happen again.

One thing that made this thing hell is how the on_fail just jumped all over the place from function to function which combined will all the currying had me scrambling to understand what is going on.

Since it is getting late I do not feel like doing anything else, but tomorrow and for the rest of the day, I am going to take some time off to see if there is a more sensible way of writing this. Partial application really ruined my day today.