Done.
This conversion so that on_fail and ret are always in the last place was actually pretty easy. It is quite remarkable.

I actually remember thinking that I should do this in the morning and deciding it would be too 'complicated'. This actually simplifies things radically. I really was not in my right state of mind in the morning.

I guess things like this happen when you are new to things.

With this bit of knowledge I think CPS should be a lot more usable to me from here on out.

At any rate, I want to continue using it until I reach an answer. I'll get better at it, I am sure.