"Now that I've worked on the language for two months, Spiral's form is starting to take shape. It seems like it will be a language specialized for metaprogramming which sounds great, but they are some tradeoffs. Compared to Haskell and F#, it will be a strictly worse term level language due to the lack of global typechecking and related to that, I won't be able to put in algebraic datatypes because it is not capable of unification due to all the dependent typing.
But on the other hand, it will be an incredibly efficient language compared to the two, and the compiler I am writing here could be written in it using just pure functions and pattern matching on open types.

In the end, given that I've decided to spend so much time making it, it is obvious that there is a great need for a language like this. Automatic differentiation in particular cannot be done efficiently in F# right now.

And compared to the other languages, Spiral's nature makes it greatly suitable as a glue language."

The above is an excerpt from my journal that describes how I feel about the future of this language.

I did a lot of improving of base cuda module and improved the error messages for the pattern matcher a tad.

I'll finish the launchers today and then I will be able to move into the library proper. Let me checkpoint it here as I need to go outside for a bit.