I am done with the 3 launcher functions. I did a lot of work to smooth things out today.
I did a lot of great work today...but looking at the 3 launcher functions it feels like there is a huge amount of wasted opportunity for abstraction that I would not have to suffer through if F# was better.

I think I will call it a day here. Before I do anything else, I need to figure out what the next steps are.

Obviously, I need to test this stuff first and foremost, but after that...

Ah, yeah, I need to make the relu and sigmoids and every other kernel, this time in a generic fashion.

So what I will be doing from here is adding those type annotations into pattern matcher, so that I do not need to pass zero and one via an argument like I am doing now in gemm.

Instead, when making Relu and Sigmoid I will just test the type of x and make ones and zeroes dependent on the type of x.